refactor: 'TestUtils' class 이름이 'PlantAlarmTestUtils'로 변경 #80
