+++ src/test/java/com/sounganization/botanify/domain/garden/service/NotificationServiceTest.java
@@ -20,7 +20,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 
 import java.util.Optional;
 
-import static com.sounganization.botanify.utils.TestUtils.createTestUser;
+import static com.sounganization.botanify.utils.PlantAlarmTestUtils.createTestUser;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
